Fixed: System.ArgumentOutOfRangeException: Value of '20480' is not valid for 'Value'. 'Value' should be between 'Minimum' and 'Maximum'. trackWatermarkBackgroundTrans value load fixed.

+++ trunk/ZScreen/Helpers/XMLSettings.cs
@@ -176,8 +176,8 @@ namespace ZSS
         // Quality
 
         public int FileFormat = 0;
-        public int ImageQuality = 90;
-        public int SwitchAfter = 512;
+        public decimal ImageQuality = 90;
+        public decimal SwitchAfter = 512;
         public int SwitchFormat = 1;
 
         //~~~~~~~~~~~~~~~~~~~~~
